+++ R/check_collinearity.R
@@ -31,6 +31,8 @@
 #'   with other terms, and tolerance values (including confidence intervals),
 #'   where `tolerance = 1/vif`.
 #'
+#' @seealso [`see::plot.see_check_collinearity()`] for options to customize the plot.
+#'
 #' @section Multicollinearity:
 #' Multicollinearity should not be confused with a raw strong correlation
 #' between predictors. What matters is the association between one or more
@@ -143,7 +145,6 @@ multicollinearity <- check_collinearity
 #' @rdname check_collinearity
 #' @export
 check_collinearity.default <- function(x, ci = 0.95, verbose = TRUE, ...) {
-  # check for valid input
   .is_model_valid(x)
   .check_collinearity(x, component = "conditional", ci = ci, verbose = verbose)
 }