+++ R/check_model.R
@@ -9,6 +9,7 @@
 #'
 #' @param x A model object.
 #' @param dot_size,line_size Size of line and dot-geoms.
+#' @param base_size,title_size,axis_title_size Base font size for axis and plot titles.
 #' @param panel Logical, if `TRUE`, plots are arranged as panels; else,
 #' single plots for each diagnostic are returned.
 #' @param check Character vector, indicating which checks for should be performed
@@ -183,19 +184,22 @@ check_model <- function(x, ...) {
 #' @rdname check_model
 #' @export
 check_model.default <- function(x,
-                                dot_size = 2,
-                                line_size = 0.8,
                                 panel = TRUE,
                                 check = "all",
+                                detrend = TRUE,
+                                bandwidth = "nrd",
+                                type = "density",
+                                residual_type = NULL,
+                                show_dots = NULL,
+                                dot_size = 2,
+                                line_size = 0.8,
+                                title_size = 12,
+                                axis_title_size = base_size,
+                                base_size = 10,
                                 alpha = 0.2,
                                 dot_alpha = 0.8,
                                 colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
                                 theme = "see::theme_lucid",
-                                detrend = TRUE,
-                                show_dots = NULL,
-                                bandwidth = "nrd",
-                                type = "density",
-                                residual_type = NULL,
                                 verbose = FALSE,
                                 ...) {
   # check model formula
@@ -209,6 +213,15 @@ check_model.default <- function(x,
   if (is.null(residual_type)) {
     residual_type <- ifelse(minfo$is_linear && !minfo$is_gam, "normal", "simulated")
   }
+
+  # catch models/families not supported by DHARMa - we need to add more
+  # exceptions here as they appear, but for now, `check_model()` also
+  # automatically falls back to normal Q-Q plot for all models not supported
+  # by DHARMa
+  if (minfo$family %in% c("quasipoisson", "quasibinomial")) {
+    residual_type <- "normal"
+  }
+
   # set default for detrend
   if (missing(detrend)) {
     detrend <- residual_type == "normal"
@@ -238,8 +251,8 @@ check_model.default <- function(x,
   }
 
   # did Q-Q plot work with simulated residuals?
-  if (verbose && is.null(assumptions_data$QQ) && residual_type == "simulated") {
-    insight::format_warning(paste0(
+  if (is.null(assumptions_data$QQ) && residual_type == "simulated") {
+    insight::format_alert(paste0(
       "Cannot simulate residuals for models of class `",
       class(x)[1],
       "`. Please try `check_model(..., residual_type = \"normal\")` instead."
@@ -261,6 +274,9 @@ check_model.default <- function(x,
   attr(assumptions_data, "panel") <- panel
   attr(assumptions_data, "dot_size") <- dot_size
   attr(assumptions_data, "line_size") <- line_size
+  attr(assumptions_data, "base_size") <- base_size
+  attr(assumptions_data, "axis_title_size") <- axis_title_size
+  attr(assumptions_data, "title_size") <- title_size
   attr(assumptions_data, "check") <- check
   attr(assumptions_data, "alpha") <- alpha
   attr(assumptions_data, "dot_alpha") <- dot_alpha
@@ -300,20 +316,23 @@ plot.check_model <- function(x, ...) {
 
 #' @export
 check_model.stanreg <- function(x,
-                                dot_size = 2,
-                                line_size = 0.8,
                                 panel = TRUE,
                                 check = "all",
+                                detrend = TRUE,
+                                bandwidth = "nrd",
+                                type = "density",
+                                residual_type = NULL,
+                                show_dots = NULL,
+                                dot_size = 2,
+                                line_size = 0.8,
+                                title_size = 12,
+                                axis_title_size = base_size,
+                                base_size = 10,
                                 alpha = 0.2,
                                 dot_alpha = 0.8,
                                 colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
                                 theme = "see::theme_lucid",
-                                detrend = TRUE,
-                                show_dots = NULL,
-                                bandwidth = "nrd",
-                                type = "density",
-                                residual_type = NULL,
-                                verbose = TRUE,
+                                verbose = FALSE,
                                 ...) {
   check_model(bayestestR::bayesian_as_frequentist(x),
     dot_size = dot_size,
@@ -324,6 +343,8 @@ check_model.stanreg <- function(x,
     dot_alpha = dot_alpha,
     colors = colors,
     theme = theme,
+    base_size = base_size,
+    axis_title_size = axis_title_size,
     detrend = detrend,
     show_dots = show_dots,
     bandwidth = bandwidth,
@@ -341,20 +362,23 @@ check_model.brmsfit <- check_model.stanreg
 
 #' @export
 check_model.model_fit <- function(x,
-                                  dot_size = 2,
-                                  line_size = 0.8,
                                   panel = TRUE,
                                   check = "all",
+                                  detrend = TRUE,
+                                  bandwidth = "nrd",
+                                  type = "density",
+                                  residual_type = NULL,
+                                  show_dots = NULL,
+                                  dot_size = 2,
+                                  line_size = 0.8,
+                                  title_size = 12,
+                                  axis_title_size = base_size,
+                                  base_size = 10,
                                   alpha = 0.2,
                                   dot_alpha = 0.8,
                                   colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
                                   theme = "see::theme_lucid",
-                                  detrend = TRUE,
-                                  show_dots = NULL,
-                                  bandwidth = "nrd",
-                                  type = "density",
-                                  residual_type = NULL,
-                                  verbose = TRUE,
+                                  verbose = FALSE,
                                   ...) {
   check_model(
     x$fit,
@@ -363,9 +387,11 @@ check_model.model_fit <- function(x,
     panel = panel,
     check = check,
     alpha = alpha,
+    axis_title_size = axis_title_size,
     dot_alpha = dot_alpha,
     colors = colors,
     theme = theme,
+    base_size = base_size,
     detrend = detrend,
     show_dots = show_dots,
     bandwidth = bandwidth,
@@ -379,20 +405,23 @@ check_model.model_fit <- function(x,
 
 #' @export
 check_model.performance_simres <- function(x,
-                                           dot_size = 2,
-                                           line_size = 0.8,
                                            panel = TRUE,
                                            check = "all",
+                                           detrend = TRUE,
+                                           bandwidth = "nrd",
+                                           type = "density",
+                                           residual_type = NULL,
+                                           show_dots = NULL,
+                                           dot_size = 2,
+                                           line_size = 0.8,
+                                           title_size = 12,
+                                           axis_title_size = base_size,
+                                           base_size = 10,
                                            alpha = 0.2,
                                            dot_alpha = 0.8,
                                            colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
                                            theme = "see::theme_lucid",
-                                           detrend = TRUE,
-                                           show_dots = NULL,
-                                           bandwidth = "nrd",
-                                           type = "density",
-                                           residual_type = NULL,
-                                           verbose = TRUE,
+                                           verbose = FALSE,
                                            ...) {
   check_model(
     x$fittedModel,
@@ -402,8 +431,10 @@ check_model.performance_simres <- function(x,
     check = check,
     alpha = alpha,
     dot_alpha = dot_alpha,
+    axis_title_size = axis_title_size,
     colors = colors,
     theme = theme,
+    base_size = base_size,
     detrend = detrend,
     show_dots = show_dots,
     bandwidth = bandwidth,